requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
Click==7.0
cryptography==2.4.2
Flask==1.0.2
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
pycparser==2.19
PyJWT==1.7.1
six==1.12.0
SQLAlchemy==1.2.15
uuid==1.30
Werkzeug==0.14.1
